% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatInput.R
\name{FormatTrack}
\alias{FormatTrack}
\title{Prepare Input for Creating Coverage Plot.}
\usage{
FormatTrack(
  data,
  region = "chr14:21,677,306-21,737,601",
  gtf.gr = NULL,
  gene.name = "HNRNPC",
  gene.name.type = c("gene_name", "gene_id"),
  extend = 2000
)
}
\arguments{
\item{data}{Track dataframe loaded by \code{\link{LoadTrackFile}}.}

\item{region}{Region used to create coverage plot, eg: chr14:21,677,306-21,737,601 or chr14:21,677,306.
Default: NULL.}

\item{gtf.gr}{Granges object of GTF, created with \code{\link{import.gff}}. Default: NULL.}

\item{gene.name}{The name of gene. Default: HNRNPC.}

\item{gene.name.type}{Gene name type (filed of \code{gtf.gr}), chosen from gene_name and gene_id.
Default: gene_name.}

\item{extend}{Extend length of \code{region}. Default: 2000.}
}
\value{
A dataframe.
}
\description{
Prepare Input for Creating Coverage Plot.
}
