% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geom_arrows}
\alias{geom_arrows}
\title{Plot genomic features as arrows.}
\usage{
geom_arrows(data, color, line_width, arrow_size, arrow_angle, arrow_type)
}
\arguments{
\item{data}{data frame describing arrow position, with columns
start, end, group, and a custom 'color' column}

\item{color}{name of the color column in the data frame}

\item{line_width}{line_width of the (arrow) segment}

\item{arrow_size}{size of the arrow}

\item{arrow_angle}{angle of the arrow in degrees}

\item{arrow_type}{type of arrow, either 'open' or 'closed'}
}
\value{
A geom layer for ggplot2 objects.
}
\description{
This function is a variation of geom_segment to plot (gene) features
as arrows. Mainly meant for internal use, not to be called directly.
}
