% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cnv.R
\name{geom_cnv}
\alias{geom_cnv}
\title{Add CNV Annotation to Coverage Plot.}
\usage{
geom_cnv(
  cnv.df,
  bin.col = 3,
  cn.col = 4,
  ref.cn = 2,
  bin.point.color = "grey",
  bin.point.alpha = 0.6,
  cn.line.color = "red",
  ref.line.color = "black",
  plot.space = 0.1,
  plot.height = 0.2
)
}
\arguments{
\item{cnv.df}{Dataframe contains copy number information, should contains at least
three columns (chr, position, bin value, copy number).}

\item{bin.col}{Column index for bin value (point). Default: 3.}

\item{cn.col}{Column index for copy number (line). Default: 4.}

\item{ref.cn}{Reference copy number (ploidy). Default: 2.}

\item{bin.point.color}{Point color of bin value. Default: "grey".}

\item{bin.point.alpha}{Point alpha of bin value. Default: 0.6.}

\item{cn.line.color}{Line color of copy number. Default: "red".}

\item{ref.line.color}{Line color of reference copy number (ploidy). Default: "black".}

\item{plot.space}{Top and bottom margin. Default: 0.1.}

\item{plot.height}{The relative height of contact map to coverage plot. Default: 0.2.}
}
\value{
Plot.
}
\description{
Add CNV Annotation to Coverage Plot.
}
\examples{
\dontrun{
  library("BSgenome.Hsapiens.UCSC.hg19")

  # load track data
  track_file <-
    system.file("extdata", "DNA-seq", "SRR054616.bw", package = "ggcoverage")
  track_df <- LoadTrackFile(
    track.file = track_file,
    format = "bw",
    region = "4:1-160000000"
  )
  track_df$seqnames <- paste0("chr", track_df$seqnames)

  # read CNV data
  cnv_file <-
    system.file("extdata", "DNA-seq", "SRR054616_copynumber.txt", package = "ggcoverage")
  cnv_df <- read.table(file = cnv_file, sep = "\t", header = TRUE)

  # plot coverage, GC content, CNV
  basic_coverage <- ggcoverage(
    data = track_df,
    color = "grey",
    mark.region = NULL,
    range.position = "out"
  )

  basic_coverage +
    geom_gc(bs.fa.seq = BSgenome.Hsapiens.UCSC.hg19) +
    geom_cnv(cnv.df = cnv_df, bin.col = 3, cn.col = 4)
}
}
