% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_feature.R
\name{geom_feature}
\alias{geom_feature}
\title{Add Feature Annotation to Coverage Plot.}
\usage{
geom_feature(
  feature.file = NULL,
  feature.df = NULL,
  feature.color = "black",
  feature.size = 5,
  plot.space = 0.1,
  plot.height = 0.1
)
}
\arguments{
\item{feature.file}{The path to feature file, should contain four columns. Default: NULL.}

\item{feature.df}{The dataframe contains feature informatin, should contain four columns. Default: NULL.}

\item{feature.color}{Feature color. Default: black.}

\item{feature.size}{The line size of peak. Default: 5.}

\item{plot.space}{Top and bottom margin. Default: 0.1.}

\item{plot.height}{The relative height of peak annotation to coverage plot. Default: 0.2.}
}
\value{
Plot.
}
\description{
Add Feature Annotation to Coverage Plot.
}
\examples{
# library(ggcoverage)
# coverage.file <- system.file(
#   "extdata", "Proteomics", "MS_BSA_coverage.xlsx", package = "ggcoverage"
# )
# fasta.file <- system.file(
#   "extdata", "Proteomics", "MS_BSA_coverage.fasta", package = "ggcoverage"
# )
# protein.id = "sp|P02769|ALBU_BOVIN"
# protein.coverage = ggprotein(
#   coverage.file = coverage.file,
#   fasta.file = fasta.file,
#   protein.id = protein.id
# )
# feature.df = data.frame(
#   ProteinID = protein.id,
#   start = c(1, 19, 25),
#   end = c(18, 24, 607),
#   Type = c("Signal", "Propeptide", "Chain")
# )
# protein.coverage +
#   geom_feature(
#     feature.df = feature.df,
#     feature.color = c("#4d81be","#173b5e","#6a521d")
#   )
}
