% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gene.R
\name{geom_gene}
\alias{geom_gene}
\title{Add Gene Annotation to Coverage Plot.}
\usage{
geom_gene(
  gtf.gr,
  overlap.gene.gap = 0.1,
  overlap.style = "loose",
  gene.size = 1,
  utr.size = 2,
  exon.size = 3,
  arrow.size = 1.5,
  arrow.gap = NULL,
  arrow.num = 50,
  color.by = "strand",
  fill.color = c(`-` = "cornflowerblue", `+` = "darkolivegreen3"),
  show.utr = FALSE,
  label.size = 3,
  label.vjust = 2,
  plot.space = 0.1,
  plot.height = 0.2
)
}
\arguments{
\item{gtf.gr}{Granges object of GTF, created with \code{\link{import.gff}}. Default: NULL.}

\item{overlap.gene.gap}{The gap between gene groups. Default: 0.1.}

\item{overlap.style}{The style of gene groups, choose from loose (each gene occupies single line)
and tight (place non-overlap genes in one line). Default: loose.}

\item{gene.size}{The line size of gene. Default: 1.}

\item{utr.size}{The line size of UTR. Default: 2.}

\item{exon.size}{The line size of exon. Default: 3.}

\item{arrow.size}{The line size of arrow. Default: 1.5.}

\item{arrow.gap}{The gap distance between intermittent arrows. Default: NULL.
Set arrow.num and arrow.gap to NULL to suppress intermittent arrows.}

\item{arrow.num}{Total number of intermittent arrows over whole region. Default: 50.
Set arrow.num and arrow.gap to NULL to suppress intermittent arrows.}

\item{color.by}{Color the line by. Default: strand.}

\item{fill.color}{Color used for \code{color.by}.
Default: blue for - (minus strand), green for + (plus strand).}

\item{show.utr}{Logical value, whether to show UTR. Default: TRUE.}

\item{label.size}{The size of gene label. Default: 3.}

\item{label.vjust}{The vjust of gene label. Default: 2.}

\item{plot.space}{Top and bottom margin. Default: 0.1.}

\item{plot.height}{The relative height of gene annotation to coverage plot. Default: 0.2.}
}
\value{
Plot.
}
\description{
Add Gene Annotation to Coverage Plot.
}
\examples{
library(ggcoverage)
library(utils)
library(rtracklayer)

# load metadata
meta_file <- system.file("extdata", "RNA-seq", "meta_info.csv", package = "ggcoverage")
sample_meta <- read.csv(meta_file)

# track folder
track_folder <- system.file("extdata", "RNA-seq", package = "ggcoverage")

# load bigwig file
track_df <- LoadTrackFile(
  track.folder = track_folder,
  format = "bw",
  region = "chr14:21,677,306-21,737,601",
  extend = 2000,
  meta.info = sample_meta
)

# load GTF file
gtf_file <- system.file("extdata", "used_hg19.gtf", package = "ggcoverage")
gtf_gr <- rtracklayer::import.gff(con = gtf_file, format = "gtf")

# plot coverage and gene annotation
basic.coverage <- ggcoverage(data = track_df, range.position = "out")
basic.coverage +
  geom_gene(gtf.gr = gtf_gr)
}
