% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ideogram.R
\name{geom_ideogram}
\alias{geom_ideogram}
\title{Add Ideogram Annotation to Coverage Plot.}
\usage{
geom_ideogram(
  genome = "hg19",
  mark.color = "red",
  mark.alpha = 0.7,
  mark.line.size = 1,
  add.shadow = TRUE,
  shadow.color = "grey",
  shadow.alpha = 0.7,
  shadow.line.size = 1,
  highlight.centromere = FALSE,
  highlight.color = "green",
  highlight.alpha = 0.7,
  highlight.line.size = 1,
  highlight.shadow.color = "black",
  highlight.shadow.alpha = 0.7,
  highlight.shadow.line.size = 1,
  plot.space = 0.1,
  plot.height = 0.1
)
}
\arguments{
\item{genome}{Single specie names, which must be one of the result from \code{ucscGenomes()$db}.
If missing, will invoke a menu for users to choose from. Default: hg19.}

\item{mark.color}{The color to mark plot region on ideogram. Default: "red".}

\item{mark.alpha}{The alpha to mark plot region on ideogram. Default: 0.7.}

\item{mark.line.size}{The line size to mark plot region on ideogram. Default: 1.}

\item{add.shadow}{Logical value, whether to add shadow polygon. Default: TRUE.}

\item{shadow.color}{The color to fill shadow polygon. Default: "grey".}

\item{shadow.alpha}{The alpha of shadow polygon. Default: 0.7.}

\item{shadow.line.size}{The line size of shadow polygon. Default: 1.}

\item{highlight.centromere}{Logical value, whether to highlight centromere region. Default: FALSE.}

\item{highlight.color}{The color to mark centromere region on ideogram. Default: "green".}

\item{highlight.alpha}{The alpha to mark centromere region on ideogram. Default: 0.7.}

\item{highlight.line.size}{The line size to mark centromere region on ideogram. Default: 1.}

\item{highlight.shadow.color}{The color to fill centromere shadow polygon. Default: "black".}

\item{highlight.shadow.alpha}{The alpha of centromere shadow polygon. Default: 0.7.}

\item{highlight.shadow.line.size}{The line size of centromere shadow polygon. Default: 1.}

\item{plot.space}{Top and bottom margin. Default: 0.1.}

\item{plot.height}{The relative height of ideogram annotation to coverage plot. Default: 0.2.}
}
\value{
Plot.
}
\description{
Add Ideogram Annotation to Coverage Plot.
}
\examples{
\dontrun{
# note that you need to have package 'ggbio' installed
library(ggbio)

# load metadata
meta_file <-
  system.file("extdata", "RNA-seq", "meta_info.csv", package = "ggcoverage")
sample_meta <- read.csv(meta_file)

# track folder
track_folder <-
  system.file("extdata", "RNA-seq", package = "ggcoverage")
# load bigwig file
track_df <- LoadTrackFile(
  track.folder = track_folder,
  format = "bw",
  region = "chr14:21,677,306-21,737,601",
  extend = 2000,
  meta.info = sample_meta
)

# gene annotation
gtf_file <-
  system.file("extdata", "used_hg19.gtf", package = "ggcoverage")
gtf_gr <- rtracklayer::import.gff(con = gtf_file, format = "gtf")

# coverage plot + ideogram
basic_coverage <- ggcoverage(
  data = track_df,
  plot.type = "facet",
  range.position = "in",
  facet.y.scale = "fixed"
)

basic_coverage +
  geom_gene(gtf.gr = gtf_gr) +
  geom_ideogram(genome = "hg19", plot.space = 0)
}

}
