% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_link.R
\name{geom_link}
\alias{geom_link}
\title{Add Links to Coverage Plot.}
\usage{
geom_link(
  link.file,
  file.type = "bedpe",
  score.col = NULL,
  score.threshold = NULL,
  score.color = c("blue", "grey", "red"),
  scale.range = 10,
  plot.space = 0.1,
  plot.height = 0.2,
  show.rect = FALSE
)
}
\arguments{
\item{link.file}{File contains region link information.}

\item{file.type}{The type of \code{link.file}, choose from bedpe, pairs. Default: bedpe.}

\item{score.col}{Column index contains score information, used when \code{file.type} is bedpe. Default: NULL.}

\item{score.threshold}{The score threshold, used when \code{score.col} is not NULL. Default: NULL.}

\item{score.color}{The score color vector. The length should be three, the first represents the lowest score, the second represents the
middle score, the third represents the highest score. Default: c("blue", "grey", "red").}

\item{scale.range}{Scale the height of link according to width, should be greater than or equal to 1 (not scale). Default: 10.}

\item{plot.space}{Top and bottom margin. Default: 0.1.}

\item{plot.height}{The relative height of link to coverage plot. Default: 0.2.}

\item{show.rect}{Logical value, whether to add rect border to the plot. Default: FALSE.}
}
\value{
Plot.
}
\description{
Add Links to Coverage Plot.
}
\examples{
library(ggcoverage)
# create test dataframe (random)
df <- data.frame(
  seqnames = "chr9", start = seq(from = 4000000, to = 5999000, by = 1000),
  end = seq(from = 4001000, to = 6000000, by = 1000), score = sample(1:100, 2000, replace = TRUE),
  Type = "Example", Group = "Example"
)
# create plot
ggcoverage(
  data = df, color = "grey", region = "chr9:4000000-6000000",
  mark.region = NULL, range.position = "out"
) +
  geom_link(link.file = link.file, file.type = "bedpe", show.rect = TRUE)
}
\references{
\url{https://stuartlab.org/signac/articles/cicero.html}
}
