% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_peak.R
\name{geom_peak}
\alias{geom_peak}
\title{Add Peak Annotation to Coverage Plot.}
\usage{
geom_peak(
  bed.file = NULL,
  peak.df = NULL,
  peak.color = NULL,
  peak.size = 5,
  color.by = NULL,
  plot.space = 0.1,
  plot.height = 0.1
)
}
\arguments{
\item{bed.file}{The path to consensus peaks file. Default: NULL.}

\item{peak.df}{The dataframe contains consensus peaks. Default: NULL.}

\item{peak.color}{Peak colors. Default: NULL.}

\item{peak.size}{The line size of peak. Default: 5.}

\item{color.by}{Name of optional column in bed file/data frame which is used for coloring. Default: NULL.}

\item{plot.space}{Top and bottom margin. Default: 0.1.}

\item{plot.height}{The relative height of peak annotation to coverage plot. Default: 0.1.}
}
\value{
Plot.
}
\description{
Add Peak Annotation to Coverage Plot.
}
\examples{
# load metadata
sample_meta <- data.frame(
  SampleName = c(
    "Chr18_MCF7_ER_1",
    "Chr18_MCF7_ER_2",
    "Chr18_MCF7_ER_3",
    "Chr18_MCF7_input"
  ),
  Type = c("MCF7_ER_1", "MCF7_ER_2", "MCF7_ER_3", "MCF7_input"),
  Group = c("IP", "IP", "IP", "Input")
)

# import coverage track
track_folder <- system.file("extdata", "ChIP-seq", package = "ggcoverage")
track_df <- LoadTrackFile(
  track.folder = track_folder,
  format = "bw",
  region = "chr18:76822285-76900000",
  meta.info = sample_meta
)

# create mock peak file
df_peaks <- data.frame(
  seqnames = c("chr18", "chr18", "chr18"),
  start = c(76822533, 76846900, 76880000),
  end = c(76836900, 76860000, 76887000),
  score = c(4, 6, 13)
)

# plot with default color
ggcoverage(data = track_df) +
  geom_peak(peak.df = df_peaks, peak.size = 3)

# plot with color by 'score' variable
ggcoverage(data = track_df) +
  geom_peak(peak.df = df_peaks, peak.size = 3, color.by = "score")

# plot with color by 'score' variable and custom color scale
ggcoverage(data = track_df) +
  geom_peak(peak.df = df_peaks, peak.size = 3, color.by = "score", peak.color = rainbow(5))

}
