% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetPlotData}
\alias{GetPlotData}
\title{Plot Object Generated by ggcoverage.}
\usage{
GetPlotData(plot, layer.num = 1)
}
\arguments{
\item{plot}{Plot.}

\item{layer.num}{The number of layers. Default: 1.}
}
\value{
A dataframe.
}
\description{
Plot Object Generated by ggcoverage.
}
\examples{
# import track data
meta.file <- system.file("extdata", "RNA-seq", "meta_info.csv", package = "ggcoverage")
sample.meta <- read.csv(meta.file)
track.folder <- system.file("extdata", "RNA-seq", package = "ggcoverage")

track.df <- LoadTrackFile(
  track.folder = track.folder,
  format = "bw",
  region = "chr14:21,677,306-21,737,601",
  extend = 2000,
  meta.info = sample.meta
)

# plot tracks
plot_coverage <- ggcoverage(
  data = track.df,
  range.position = "out",
  show.mark.label = TRUE
)

# extract data again
plot_data <- GetPlotData(plot = plot_coverage, layer.num = 1)
head(plot_data)

}
