% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggprotein.R
\name{ggprotein}
\alias{ggprotein}
\title{Create Mass Spectrometry Protein Coverage Plot.}
\usage{
ggprotein(
  coverage.df,
  fasta.file,
  protein.id,
  XCorr.threshold = 2,
  confidence = "High",
  contaminant = NULL,
  remove.na = TRUE,
  color = "grey",
  mark.bare = TRUE,
  mark.color = "red",
  mark.alpha = 0.5,
  show.table = TRUE,
  table.position = c("top_right", "top_left", "bottom_left", "bottom_right"),
  table.size = 10,
  table.color = "black",
  range.size = 3,
  range.position = c("in", "out")
)
}
\arguments{
\item{coverage.df}{Protein coverage, for example output from Proteome Discoverer.}

\item{fasta.file}{Input reference protein fasta file.}

\item{protein.id}{The protein ID of exported coverage file. This should be unique and in \code{fasta.file}.}

\item{XCorr.threshold}{The cross-correlation threshold. Default: 2.}

\item{confidence}{The confidence level. Default: High.}

\item{contaminant}{Whether to remove contaminant peptides. Default: NULL (not remove).}

\item{remove.na}{Logical value, whether to remove NA value in Abundance column. Default: TRUE.}

\item{color}{The fill color of coverage plot. Default: grey.}

\item{mark.bare}{Logical value, whether to mark region where Abundance is zero or NA. Default: TRUE.}

\item{mark.color}{The color used for the marked region. Default: red.}

\item{mark.alpha}{The transparency used for the marked region. Default: 0.5.}

\item{show.table}{Logical value, whether to show coverage summary table. Default: TRUE.}

\item{table.position}{The position of the coverage summary table, choose from right_top, left_top, left_bottom, right_bottom.
Default: right_top.}

\item{table.size}{The font size of coverage summary table. Default: 4.}

\item{table.color}{The font color of coverage summary table. Default: black.}

\item{range.size}{The label size of range text, used when \code{range.position} is in. Default: 3.}

\item{range.position}{The position of y axis range, chosen from in (move y axis in the plot) and
out (normal y axis). Default: in.}
}
\value{
A ggplot2 object.
}
\description{
Create Mass Spectrometry Protein Coverage Plot.
}
\examples{
\donttest{
  if (requireNamespace("openxlsx", quietly = TRUE)) {
    library(ggplot2)
    library(openxlsx)

    # import coverage dataframe with function from openxlsx
    coverage.file <- system.file(
      "extdata", "Proteomics", "MS_BSA_coverage.xlsx",
      package = "ggcoverage"
    )
    coverage.df <- read.xlsx(coverage.file)
    head(coverage.df)

    # get fasta file
    fasta.file <- system.file(
      "extdata", "Proteomics", "MS_BSA_coverage.fasta",
      package = "ggcoverage"
    )

    protein.id <- "sp|P02769|ALBU_BOVIN"
    ggprotein(
      coverage.df = coverage.df,
      fasta.file = fasta.file,
      protein.id = protein.id
    )
  }
}
}
