% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggcoverage.R
\name{theme_aa}
\alias{theme_aa}
\title{Theme for geom_base with Amino Acid.}
\usage{
theme_aa(margin.len, fill.color)
}
\arguments{
\item{margin.len}{Top and bottom margin.}

\item{fill.color}{Fill color.}
}
\value{
List of layers.
}
\description{
Theme for geom_base with Amino Acid.
}
