% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggcoverage.R
\name{theme_base}
\alias{theme_base}
\title{Theme for geom_base.}
\usage{
theme_base(margin.len, fill.color)
}
\arguments{
\item{margin.len}{Top and bottom margin.}

\item{fill.color}{Fill color.}
}
\value{
List of layers.
}
\description{
Theme for geom_base.
}
