% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggcoverage.R
\name{theme_base2}
\alias{theme_base2}
\title{Theme for geom_base without margin.}
\usage{
theme_base2(fill.color)
}
\arguments{
\item{fill.color}{Fill color.}
}
\value{
List of layers.
}
\description{
Theme for geom_base without margin.
}
