% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggcoverage.R
\name{theme_coverage}
\alias{theme_coverage}
\title{Theme for geom_coverage.}
\usage{
theme_coverage(space = 0.2, x.range)
}
\arguments{
\item{space}{The space between facets. Default: 0.2.}

\item{x.range}{X axis ranges.}
}
\value{
List of layers.
}
\description{
Theme for geom_coverage.
}
