% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggcoverage.R
\name{theme_feature}
\alias{theme_feature}
\title{Theme for geom_feature.}
\usage{
theme_feature(margin.len, x.range, fill.color)
}
\arguments{
\item{margin.len}{Top and bottom margin.}

\item{x.range}{X axis ranges.}

\item{fill.color}{Fill color for different feature type.}
}
\value{
List of layers.
}
\description{
Theme for geom_feature.
}
