% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggcoverage.R
\name{theme_gc}
\alias{theme_gc}
\title{Theme for geom_gc.}
\usage{
theme_gc(x.range, margin.len)
}
\arguments{
\item{x.range}{X axis ranges.}

\item{margin.len}{Top and bottom margin.}
}
\value{
List of layers.
}
\description{
Theme for geom_gc.
}
