% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggcoverage.R
\name{theme_gene}
\alias{theme_gene}
\title{Theme for geom_gene.}
\usage{
theme_gene(overlap.gene.gap, group.num, fill.color, x.range, margin.len)
}
\arguments{
\item{overlap.gene.gap}{The gap between gene groups.}

\item{group.num}{The number of groups.}

\item{fill.color}{Fill color.}

\item{x.range}{X axis ranges.}

\item{margin.len}{Top and bottom margin.}
}
\value{
List of layers.
}
\description{
Theme for geom_gene.
}
