% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggcoverage.R
\name{theme_peak}
\alias{theme_peak}
\title{Theme for geom_peak.}
\usage{
theme_peak(margin.len, x.range)
}
\arguments{
\item{margin.len}{Top and bottom margin.}

\item{x.range}{X axis ranges.}
}
\value{
List of layers.
}
\description{
Theme for geom_peak.
}
