% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggcoverage.R
\name{theme_tad2}
\alias{theme_tad2}
\title{Theme for geom_tad2.}
\usage{
theme_tad2(x.range, margin.len, show.rect)
}
\arguments{
\item{x.range}{X axis ranges.}

\item{margin.len}{Top and bottom margin.}

\item{show.rect}{Logical value, whether to add rect border to the plot.}
}
\value{
List of layers.
}
\description{
Theme for geom_tad2.
}
