% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggcoverage.R
\name{theme_transcript}
\alias{theme_transcript}
\title{Theme for geom_transcript.}
\usage{
theme_transcript(overlap.tx.gap, group.num, fill.color, x.range, margin.len)
}
\arguments{
\item{overlap.tx.gap}{The gap between transcript groups.}

\item{group.num}{The number of groups.}

\item{fill.color}{Fill color.}

\item{x.range}{X axis ranges.}

\item{margin.len}{Top and bottom margin.}
}
\value{
List of layers.
}
\description{
Theme for geom_transcript.
}
