% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gene.R
\name{geom_gene}
\alias{geom_gene}
\title{Add Gene Annotation to Coverage Plot.}
\usage{
geom_gene(
  gtf.gr,
  overlap.gene.gap = 0.1,
  overlap.style = "loose",
  gene.size = 1,
  utr.size = 2,
  exon.size = 3,
  arrow.angle = 35,
  arrow.length = 1.5,
  arrow.type = "open",
  color.by = "strand",
  arrow.gap = NULL,
  arrow.num = 50,
  arrow.size.im = 0.5,
  arrow.length.im = 1.5,
  arrow.type.im = "closed",
  color.by.im = NULL,
  fill.color = c(`-` = "cornflowerblue", `+` = "darkolivegreen3"),
  show.utr = FALSE,
  label.size = 3,
  label.vjust = 2,
  plot.space = 0.1,
  plot.height = 0.2
)
}
\arguments{
\item{gtf.gr}{Granges object of GTF, created with \code{\link{import.gff}}.
Default: NULL.}

\item{overlap.gene.gap}{The gap between gene groups. Default: 0.1.}

\item{overlap.style}{The style of gene groups, choose from loose (each gene
occupies single line) and tight (place non-overlap genes in one line).
Default: loose.}

\item{gene.size}{Line width of genes. Default: 1.}

\item{utr.size}{Line width of UTRs. Default: 2.}

\item{exon.size}{Line width of exons. Default: 3.}

\item{arrow.angle}{Angle of the arrow head. Default 35°}

\item{arrow.length}{Length of arrows. Default: 1.5}

\item{arrow.type}{Whether to draw "closed" or "open" (default) arrow heads}

\item{color.by}{Color the lines/arrows by variable. Default: "strand".}

\item{arrow.gap}{The gap distance between intermittent arrows. Default: NULL.
Set arrow.num and arrow.gap to NULL to suppress intermittent arrows.}

\item{arrow.num}{Total number of intermittent arrows over whole region.
Default: 50. Set arrow.num and arrow.gap to NULL to suppress intermittent
arrows.}

\item{arrow.size.im}{Line width of intermittent arrows. Default: 0.5}

\item{arrow.length.im}{Length of intermittent arrows. Default: 1.5}

\item{arrow.type.im}{Whether to draw "closed" (default) or "open" heads for
intermittent arrows}

\item{color.by.im}{Color the intermittent arrows by variable. Default: NULL
(draws semi-transparent, white arrows)}

\item{fill.color}{Color used for \code{color.by}.
Default: blue for - (minus strand), green for + (plus strand).}

\item{show.utr}{Logical value, whether to show UTR. Default: TRUE.}

\item{label.size}{The size of gene label. Default: 3.}

\item{label.vjust}{The vjust of gene label. Default: 2.}

\item{plot.space}{Top and bottom margin. Default: 0.1.}

\item{plot.height}{The relative height of gene annotation to coverage plot.
Default: 0.2.}
}
\value{
Plot.
}
\description{
Add Gene Annotation to Coverage Plot.
}
\examples{
library(ggcoverage)
library(utils)
library(rtracklayer)

# load metadata
meta_file <- system.file("extdata", "RNA-seq", "meta_info.csv", package = "ggcoverage")
sample_meta <- read.csv(meta_file)

# track folder
track_folder <- system.file("extdata", "RNA-seq", package = "ggcoverage")

# load bigwig file
track_df <- LoadTrackFile(
  track.folder = track_folder,
  format = "bw",
  region = "chr14:21,677,306-21,737,601",
  extend = 2000,
  meta.info = sample_meta
)

# load GTF file
gtf_file <- system.file("extdata", "used_hg19.gtf", package = "ggcoverage")
gtf_gr <- rtracklayer::import.gff(con = gtf_file, format = "gtf")

# plot coverage and gene annotation
basic_coverage <- ggcoverage(data = track_df, range.position = "out")
basic_coverage +
  geom_gene(gtf.gr = gtf_gr)

# plot with custom style
basic_coverage +
  geom_gene(
    gtf.gr = gtf_gr,
    exon.size = 2.0,
    arrow.size.im = 1.0,
    arrow.length.im = 5,
    arrow.type.im = "open",
    color.by.im = "strand",
    fill.color = c(
      "-" = "darkblue",
      "+" = "darkgreen"
    )
  )
}
