% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_peak.R
\name{geom_peak}
\alias{geom_peak}
\title{Add Peak Annotation to Coverage Plot.}
\usage{
geom_peak(
  bed.file,
  peak.color = "black",
  peak.size = 5,
  plot.space = 0.1,
  plot.height = 0.1
)
}
\arguments{
\item{bed.file}{The path to bed file.}

\item{peak.color}{Peak color. Default: black.}

\item{peak.size}{The line size of peak. Default: 5.}

\item{plot.space}{Top and bottom margin. Default: 0.1.}

\item{plot.height}{The relative height of gene annotation to coverage plot. Default: 0.2.}
}
\value{
Plot.
}
\description{
Add Peak Annotation to Coverage Plot.
}
