% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_transcript.R
\name{geom_transcript}
\alias{geom_transcript}
\title{Add Transcript Annotation to Coverage Plot.}
\usage{
geom_transcript(
  gtf.gr,
  gene.name = "HNRNPC",
  overlap.tx.gap = 0.1,
  overlap.style = "loose",
  tx.size = 1,
  utr.size = 2,
  exon.size = 3,
  arrow.angle = 35,
  arrow.length = 1.5,
  arrow.type = "open",
  color.by = "strand",
  arrow.gap = NULL,
  arrow.num = 50,
  arrow.size.im = 0.5,
  arrow.length.im = 1.5,
  arrow.type.im = "closed",
  color.by.im = NULL,
  fill.color = c(`-` = "cornflowerblue", `+` = "darkolivegreen3"),
  label.size = 3,
  label.vjust = 2,
  plot.space = 0.1,
  plot.height = 1
)
}
\arguments{
\item{gtf.gr}{Granges object of GTF, created with \code{\link[rtracklayer]{import.gff}}.
Default: NULL.}

\item{gene.name}{Gene name of all transcripts. Default: HNRNPC.}

\item{overlap.tx.gap}{The gap between transcript groups. Default: 0.1.}

\item{overlap.style}{The style of transcript groups, choose from loose (each
transcript occupies single line) and tight (place non-overlap transcripts
in one line). Default: loose.}

\item{tx.size}{Line size of transcript. Default: 1.}

\item{utr.size}{Line size of UTR. Default: 2.}

\item{exon.size}{Line size of exon. Default: 3.}

\item{arrow.angle}{Angle of the arrow head. Default 35°}

\item{arrow.length}{Length of arrows. Default: 1.5}

\item{arrow.type}{Whether to draw "closed" or "open" (default) arrow heads}

\item{color.by}{Color the line by. Default: strand.}

\item{arrow.gap}{The gap distance between intermittent arrows. Default: NULL.
Set arrow.num and arrow.gap to NULL to suppress intermittent arrows.}

\item{arrow.num}{Total number of intermittent arrows over whole region.
Default: 50. Set arrow.num and arrow.gap to NULL to suppress intermittent
arrows.}

\item{arrow.size.im}{Line width of intermittent arrows. Default: 0.5}

\item{arrow.length.im}{Length of intermittent arrows. Default: 1.5}

\item{arrow.type.im}{Whether to draw "closed" (default) or "open" heads for
intermittent arrows}

\item{color.by.im}{Color the intermittent arrows by variable. Default: NULL
(draws semi-transparent, white arrows)}

\item{fill.color}{Color used for \code{color.by}. Default: blue for - (minus
strand), green for + (plus strand).}

\item{label.size}{The size of transcript label. Default: 3.}

\item{label.vjust}{The vjust of transcript label. Default: 2.}

\item{plot.space}{Top and bottom margin. Default: 0.1.}

\item{plot.height}{The relative height of transcript annotation to coverage
plot. Default: 0.2.}
}
\value{
Plot.
}
\description{
Add Transcript Annotation to Coverage Plot.
}
\examples{
# load metadata
meta_file <- system.file("extdata", "RNA-seq", "meta_info.csv", package = "ggcoverage")
sample_meta <- read.csv(meta_file)

# track folder
track_folder <- system.file("extdata", "RNA-seq", package = "ggcoverage")

# load bigwig file
track_df <- LoadTrackFile(
  track.folder = track_folder,
  format = "bw",
  region = "chr14:21,677,306-21,737,601",
  extend = 2000,
  meta.info = sample_meta
)

# load GTF file
gtf_file <- system.file("extdata", "used_hg19.gtf", package = "ggcoverage")
gtf_gr <- rtracklayer::import.gff(con = gtf_file, format = "gtf")

# plot coverage and gene annotation
basic_coverage <- ggcoverage(data = track_df, range.position = "out")
basic_coverage +
  geom_transcript(gtf.gr = gtf_gr, label.vjust = 1.5)

}
