% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetPlotData}
\alias{GetPlotData}
\title{Plot Object Generated by ggcoverage.}
\usage{
GetPlotData(plot, layer.num = 1)
}
\arguments{
\item{plot}{Plot.}

\item{layer.num}{The number of layers. Default: 1.}
}
\value{
A dataframe.
}
\description{
Plot Object Generated by ggcoverage.
}
\examples{
# cov.plot = ggcoverage(data = track.df, color = "auto", region = "chr18:76822285-76900000",
#                       range.position = "out", mark.region=mark.region, show.mark.label = TRUE)
# plot.data = GetPlotData(plot = cov.plot, layer.num=1)
}
